MAINT: Move validator facilities
