+++ src/apps/scdetect/CMakeLists.txt
@@ -9,6 +9,7 @@ set(
     config/detector.cpp
     config/exception.cpp
     config/template_family.cpp
+    config/validators.cpp
     datamodel/ddl.cpp
     detail/sqlite.cpp
     detector/arrival.cpp
@@ -41,7 +42,6 @@ set(
     util/horizontal_components.cpp
     util/util.cpp
     util/waveform_stream_id.cpp
-    validators.cpp
     waveform.cpp
     waveformoperator.cpp
     waveformprocessor.cpp
