+++ src/apps/scdetect/config/validators.h
@@ -1,5 +1,5 @@
-#ifndef SCDETECT_APPS_SCDETECT_VALIDATORS_H_
-#define SCDETECT_APPS_SCDETECT_VALIDATORS_H_
+#ifndef SCDETECT_APPS_SCDETECT_CONFIG_VALIDATORS_H_
+#define SCDETECT_APPS_SCDETECT_CONFIG_VALIDATORS_H_
 
 #include <string>
 #include <vector>
@@ -22,4 +22,4 @@ bool validateLinkerMergingStrategy(const std::string &mergingStrategy);
 }  // namespace detect
 }  // namespace Seiscomp
 
-#endif  // SCDETECT_APPS_SCDETECT_VALIDATORS_H_
+#endif  // SCDETECT_APPS_SCDETECT_CONFIG_VALIDATORS_H_
